Completed #9 but uncovered incompatability #14
I edited ConvertNetwork in order to close #9 which was fairly straight
forward. However, although the converted network could be loaded without
error, actually executing it resulted in the error described in #14 .
It's possible that the code is compatible between MM-NEAT and this
project, and some setting/configuration is simply set wrong, but it is
also possible there is a code descrepancy. Need to find it.

3MTTest.xml was added as an example file converted from the current
version of MM-NEAT. It should be used to solve #14 but can be deleted
afterward.