The multi-experiment runner is simpler than you thought
There is code built in already to run the agents multiple times. This seems to work fine, but only reports the statistics. Also, the average score is horrible! Unless I'm running the wrong agent ... I don't know why I'm getting an average around 2000. The post-evals from MM-NEAT give averages closer to 5000 (slightly less, but this is still a big difference)
@squillyprice01