Wrong network still?
Somehow the network from MMNEAT in MsPacManEntry had the right id number but was different from the one in MMNEAT. Weird conincidence? Re-copied the network, converted it again, and compiled with Maven again. Also changed the network converting code to keep carriage returns for better formatting.

Got an average score across 10 runs of 8671! Looks good, but need to test more.