Module Color Paths Work!
@squillyprice01
All we had to do was set both modePheremone and watch to true! The tricky part was setting watch to true ... we mistakenly though that because we could see the game that all parameters associated with watching it were set correctly.

Once this was set, I did realize there were some null pointer exceptions that could occur due to not synchonizing access to certain variables, but this was easy to fix.

Incidentally, I saw the agent reach a score of 15,000 by beating all levels, but once this happened the level looped back to the first level. However, at this point the agent basically gave up and the message Reached MAX Level kept printing out. Should fix this.