Can launch evolved champions of different types
There is an example of each type of evolved network in the resources dir now. I had to change the way MyPacMan is launched to accommodate the different Parameters required to make each one behavior correctly.

Interestingly, the various preference neuron champions seem to mainly rely on a single module.

I plan to add ALL evolved champions to the resourced soon so that we can do the post evals in this project.