Adjustments for StatsRun
Not sure how I want the game to behave once the "max level" is reached. Currently, the agent returns an exit code. For stats run, we don't want to end the simulation, but doing a neutral action doesn't immediately end eval either. Need to work on this.

Also, I'm a bit worried about the Multitask performance after doing the clean sweep with MMNEAT champions, but I don't know if the score difference is simply due to the high variance.