Fixed stats freeze?
I think the freeze that occurred with the AllPostEvals test was actually deadlock (maybe?) or at least somehow related to the synchronization added for drawing the pill and ghost model. However, there is no need to update/maintain all of that drawing information when there is no visualization, so I set the MMNEAT Parameter "watch" to false when doing a stats run (in this code base, the use of "watch" has no influence on whether the visualization occurs, but it does influence behind-the-scenes data tracking for the visualization).